+++ OneBlink.SDK/models/WorkflowAttachmentUploadCredentials.cs
@@ -1,0 +1,10 @@
+namespace OneBlink.SDK.Model
+{
+    public class WorkflowAttachmentUploadCredentials
+    {
+        public string filename
+        {
+            get; set;
+        }
+    }
+}